+++ src/ScheduledTask.h
@@ -4,15 +4,15 @@
 struct ScheduledTask
 {
 	int id;
-	double begin;
-	double end;
+    int begin;
+    int end;
 
 	ScheduledTask()
 	{
 		//Empty
 	}
 
-	ScheduledTask(int _id, double _begin, double _end)
+    ScheduledTask(int _id, int _begin, int _end)
 	{
 		id = _id;
 		begin = _begin;
